Add files via upload
This is world energy consumption tableau data analytics project 
and this is the real dataset https://www.kaggle.com/datasets/pralabhpoudel/world-energy-consumption/data
i did so many changes in original dataset as there is so many blank columns and also there is no proper data 
but its a big dataset its but disturbing to work with this type of dataset but on the other hand its also intresting to work with this dataset. it contain lot for challenges but also lot of learnings also.